requirements.txt
asgiref==3.4.1
Django~=4.0.1
Pillow==8.4.0
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
djangorestframework~=3.13.1
freezegun~=1.1.0